Update RailroaderRD.cs to Ignore RailDriver when locomotive is in AI mode
Before this fix if you put your selected locomotive in AI mode (road or yard) the update method will continuously override the locomotive and the locomotive will not move per the AI's instructions. 

This commit makes the update method check for AI modes, and if in an AI mode, do not check the rail drivers current state.